since not allowing folder creation created file.
in the repository not able to create folder followed below tip to create folder.
https://intellipaat.com/community/20249/create-folder-in-github-how-to-create-folder-in-github-repository
You cannot create an empty folder and then add files to that folder, but instead, the creation of a folder must happen together with the addition of at least a single file. On Github you can do it this way:

go to the folder within which you would like to make another folder

Click on New file

on the text field for the file name, first, write the folder name you want to create

then type /. This creates a folder

you can add more folders similarly

finally, give the new file a name, (eg. .gitkeep which is conventionally used to allow git to track otherwise empty folders; not a git feature though)

finally, click Commit new file.You cannot create an empty folder and then add files to that folder, but instead, the creation of a folder must happen together with the addition of at least a single file. On Github you can do it this way:

go to the folder within which you would like to make another folder

Click on New file

on the text field for the file name, first, write the folder name you want to create

then type /. This creates a folder

you can add more folders similarly

finally, give the new file a name, (eg. .gitkeep which is conventionally used to allow git to track otherwise empty folders; not a git feature though)

finally, click Commit new file.